refactor(ui): refactor utils
