+++ web/src/components/Transaction/index.tsx
@@ -2,7 +2,8 @@ import React from 'react'
 import { Link } from 'react-router-dom'
 import { TransactionsCell, TransactionsItem, CellHash, CellHashHighLight } from './styled'
 import { parseDate } from '../../utils/date'
-import { shannonToCkb, startEndEllipsis } from '../../utils/util'
+import { shannonToCkb } from '../../utils/util'
+import { startEndEllipsis } from '../../utils/string'
 import InputOutputIcon from '../../asserts/input_arrow_output.png'
 import { InputOutput} from '../../http/response/Transaction'
